Move file functions to list_files
